desarrollo e integracion base para opciones de respuestas en item (expansion)

+++ app/src/main/java/mx/checklist/data/api/dto/RespondReq.kt
@@ -2,13 +2,14 @@ package mx.checklist.data.api.dto
 
 /**
  * Request para responder un ítem.
- * - responseStatus: requerido
+ * - responseStatus: requerido (solo "OK" o "FAIL")
  * - responseText: opcional
  * - responseNumber: opcional (Double?)
+ * - scannedBarcode: opcional
  */
 data class RespondReq(
-    val responseStatus: String, // "OK", "FAIL", "NA"
+    val responseStatus: String, // "OK", "FAIL"
     val responseText: String? = null,
-    val responseNumber: Int? = null,
+    val responseNumber: Double? = null,
     val scannedBarcode: String? = null
 )
